feat: hwLite 클래스와 webApiConnector용 클래스 로직 분리
