+++ src/class/engine.js
@@ -747,7 +747,7 @@ Entry.Engine = class Engine {
         Entry.scene.loadStartSceneSnapshot();
         Entry.Func.clearThreads();
         Entry.Utils.setVolume(1);
-        if (Entry.hwLite.status === Entry.hwLite.HardwareStatement.connected) {
+        if (Entry.hwLite.getStatus() === 'connected') {
             Entry.hwLite.setZero();
         }
         createjs.Sound.setVolume(1);
